Split up views into individual files in the views directory
