+++ app/views/login.py
@@ -1,0 +1,43 @@
+from flask import session, redirect, url_for, request, render_template, flash
+from app import app, forms
+from datetime import datetime
+import pymesync
+import re
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    status = 200
+    form = forms.LoginForm()
+
+    # If POST request and form valid, login user
+    if form.validate_on_submit():
+        username = form.username.data
+        password = form.password.data
+
+        # Login to TimeSync
+        ts = pymesync.TimeSync(baseurl=app.config['TIMESYNC_URL'],
+                               test=app.config['TESTING'])
+        token = ts.authenticate(username=username,
+                                password=password,
+                                auth_type='password')
+
+        # If regular error, tell user error
+        if 'error' in token:
+            flash(token['text'])
+            status = token['status']
+        # If pymesync error, tell user vague error
+        elif 'pymesync error' in token:
+            print token
+            return "There was an error.", 500
+        # Else success, redirect to index page
+        else:
+            session['username'] = username
+            session['token'] = token['token']
+            return form.redirect(url_for('index'))
+
+    # Else if POST request (meaning form invalid), notify user
+    elif request.method == 'POST':
+        flash("Invalid submission.")
+        status = 401
+
+    return render_template('login.html', form=form), status
