+++ tests/test_create_time.py
@@ -1,5 +1,6 @@
 import unittest
 from app import app
+from app.views import *
 from flask import url_for
 from urlparse import urlparse
