+++ tests/test_index.py
@@ -1,5 +1,6 @@
 import unittest
 from app import app
+from app.views import *
 from flask import url_for
