+++ tests/test_view_times.py
@@ -1,6 +1,6 @@
 import unittest
-from app import app
-from app import forms
+from app import app, forms
+from app.views import *
 from flask import url_for
 from urlparse import urlparse
